requirements.txt
streamlit
PyPDF2==3.0.1
python-dotenv==0.20.0
langchain-community
openai
altair<5
faiss-cpu
tiktoken
sentence_transformers==2.2.2
InstructorEmbedding
text-generation
transformers
google-search-results
numexpr
langchainhub
sentencepiece
jinja2
huggingface_hub
